Slots: Extract drag and drop logic into utils file
