+++ static/js/redux/ui/slot.jsx
@@ -19,25 +19,15 @@ import { DropTarget } from "react-dnd";
 import COLOUR_DATA from "../constants/colours";
 import { DRAG_TYPES, HALF_HOUR_HEIGHT } from "../constants/constants";
 import * as SemesterlyPropTypes from "../constants/semesterlyPropTypes";
-import { convertHalfHoursToStr, getNewSlotValues } from "./slotUtils";
-
-function convertToHalfHours(str) {
-  const start = parseInt(str.split(":")[0], 10);
-  return str.split(":")[1] === "30" ? start * 2 + 1 : start * 2;
-}
+import {
+  onCustomSlotCreateDrag,
+  onCustomSlotCreateDrop,
+  onCustomSlotUpdateDrop,
+} from "./slotUtils";
 
 const dragSlotTarget = {
   drop(props, monitor) {
-    // move it to current location on drop
-    const { timeStart, timeEnd, id } = monitor.getItem();
-
-    const slotTop = $(`#${props.id}`).offset().top;
-    // number half hours from slot start
-    const n = Math.floor((monitor.getClientOffset().y - slotTop) / HALF_HOUR_HEIGHT);
-
-    const newStartHour = convertToHalfHours(props.time_start) + n;
-    const newValues = getNewSlotValues(timeStart, timeEnd, newStartHour, props.day);
-    props.updateCustomSlot(newValues, id);
+    onCustomSlotUpdateDrop(props, monitor);
   },
 };
 
@@ -48,39 +38,12 @@ function collectDragDrop(connect) {
   };
 }
 
-let lastPreview = null;
 const createSlotTarget = {
   drop(props, monitor) {
-    // move it to current location on drop
-    let { timeStart } = monitor.getItem();
-    const { id } = monitor.getItem();
-
-    // get the time that the mouse dropped on
-    const slotTop = $(`#${props.id}`).offset().top;
-    const n = Math.floor((monitor.getClientOffset().y - slotTop) / HALF_HOUR_HEIGHT);
-    let timeEnd = convertHalfHoursToStr(convertToHalfHours(props.time_start) + n);
-
-    if (convertToHalfHours(timeStart) > convertToHalfHours(timeEnd)) {
-      [timeStart, timeEnd] = [timeEnd, timeStart];
-    }
-    props.updateCustomSlot({ time_start: timeStart, time_end: timeEnd }, id, true);
+    onCustomSlotCreateDrop(props, monitor);
   },
   hover(props, monitor) {
-    let { timeStart } = monitor.getItem();
-    const { id } = monitor.getItem();
-
-    // get the time that the mouse dropped on
-    const slotTop = $(`#${props.id}`).offset().top;
-    const n = Math.floor((monitor.getClientOffset().y - slotTop) / HALF_HOUR_HEIGHT);
-    if (n === lastPreview) {
-      return;
-    }
-    let timeEnd = convertHalfHoursToStr(convertToHalfHours(props.time_start) + n);
-    if (convertToHalfHours(timeStart) > convertToHalfHours(timeEnd)) {
-      [timeStart, timeEnd] = [timeEnd, timeStart];
-    }
-    lastPreview = n;
-    props.updateCustomSlot({ time_start: timeStart, time_end: timeEnd }, id);
+    onCustomSlotCreateDrag(props, monitor);
   },
 };
